created a main layout and stuck all navigation into it
